refactor: separate category CLI into admin (create/delete) and user (search/get)
Admin CLI keeps only write operations (create, delete).
User-facing CLI provides read operations (search, get) available to
any authenticated user, registered in the v2 CLI group.
