+++ src/ai/backend/client/cli/v2/prometheus_query_preset_category/__init__.py
@@ -1,0 +1,3 @@
+from .commands import prometheus_query_preset_category
+
+__all__ = ("prometheus_query_preset_category",)
